Fixed a few bugs in the tweaked turn system and added wall destruction
Fixed bugs which cause turns to pass when the player cancelled choosing
and/or using an item or pressing an invalid key.
Added unbreakable walls at the edge of the map.
Added a new spell called Armageddon which destroys walls in range and
deals massive damage to enemies and player. Cannot be aimed, always take
the player's position as origin, so player always take damage.