Created docs/design folders.
Design folder is for ideas that are yet to be implemented, and/or that
need to be discussed before making it to the game. Lore-related texts
(i.e scenario, dialogues, intro-text) can also go here.
Comitting/syncing text files works exactly the same as with code, so you
can make modifications to them this way (especially useful for stuff
like descriptions, or long lore related-text, such as the intro text
once I finish writing it).