Partially rewrote beginning of spellGen
Realized my first spell generation idea was flawed from a game-balance
point of view, so I decided to rewrite some stuff in the few lines I
already wrote. Turns out this caused everything to break, so had to
write a lot of not so useful lines to try and fix it.

TL;DR : Spell generation hasn't progressed since last commit.